+++ morebuiltins/__init__.py
@@ -1,4 +1,6 @@
 __version__ = "1.3.2"
+
+# this __all__ is used for documentation generation, not for imports
 __all__ = [
     "morebuiltins.utils",
     "morebuiltins.date",