+++ morebuiltins/__init__.py
@@ -1,4 +1,4 @@
-__version__ = "1.3.2"
+__version__ = "1.3.3"
 
 # this __all__ is used for documentation generation, not for imports
 __all__ = [